Applied XCode 8 recommended settings + fixed all new warnings
With XCode 8 we got many more warnings which this patch fixes, along with a bit clean up. Significant changes were made for performSelector: calls, which often gave warnings about unknown selectors and leaks (which is normal with ARC, but needed a workaround).

+++ backend/wbprivate/workbench/wb_model_file.cpp
@@ -753,7 +753,7 @@ void ModelFile::pack_zip(const std::string &zipfile, const std::string &destdir,
 #if defined(zip_uint16_t) || defined(_WIN32)
   zip_set_archive_comment(z, zip_comment.c_str(), (zip_uint16_t)zip_comment.size());
 #else
-  zip_set_archive_comment(z, zip_comment.c_str(), zip_comment.size());
+  zip_set_archive_comment(z, zip_comment.c_str(), (int)zip_comment.size());
 #endif
 
   try
